requirements.txt
appdirs==1.4.4
asgiref==3.4.1
awesome-slugify==1.6.5
black==21.7b0
click==8.0.1
colorama==0.4.4
crispy-bootstrap5==0.4
Django==3.2.5
django-ckeditor==6.1.0
django-constance==2.8.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-picklefield==3.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
python-dotenv==0.19.0
pytz==2021.1
regex==2021.7.6
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tomli==1.1.0
Unidecode==0.4.21